- Apply Unit of Work for order function because there are many operations on it and they should be wrapped in only one transaction, Unit of Work help to achieve that easily - Fixed an issue where spaces were not removed before performing a product search - Fix an issue that LinQ queries could not be translated in account management - Refactoring code (Correting error messages)

+++ OnlineStore/Areas/Admin/Pages/Order/ReceivingTypeManagement.cshtml.cs
@@ -16,6 +16,7 @@ namespace OnlineStore.Areas.Admin.Pages.Order
     {
         private readonly IReceivingTypeRepository _receivingTypeRepository;
         private readonly IOrderRepository _orderRepository;
+
         public ReceivingTypeManagementModel(
             IReceivingTypeRepository receivingTypeRepository,
             IOrderRepository orderRepository)
@@ -23,9 +24,9 @@ namespace OnlineStore.Areas.Admin.Pages.Order
             _receivingTypeRepository = receivingTypeRepository;
             _orderRepository = orderRepository;
         }
+
         public void OnGet()
         {
-
         }
 
         public IActionResult OnGetAllPaging()
@@ -70,7 +71,7 @@ namespace OnlineStore.Areas.Admin.Pages.Order
         {
             var receivingType = _receivingTypeRepository.Find(model.Id);
             _receivingTypeRepository.Delete(receivingType);
-            var orders = _orderRepository.GetSome(x => x.ReceivingTypeId == model.Id && x.IsDeleted == false);
+            var orders = _orderRepository.GetSome(x => (int)x.ReceivingTypeId == model.Id && x.IsDeleted == false);
             _orderRepository.DeleteRange(orders);
             return new OkResult();
         }
